+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change Log
 
 
+## Unreleased
+
+### Changed
+
+- Restructured code
+- Status checker uses RWMutex
+
+
 ## 0.7.0 - 2017-05-31
 
 ### Added